Move everything under one command
