--- cmd/selfupdate/check/command.go
@@ -1,61 +1,0 @@
-package check
-
-import (
-	"io"
-	"os"
-
-	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-	"github.com/spf13/afero"
-	"github.com/spf13/cobra"
-)
-
-const (
-	name = "check"
-
-	description = "Check for new versions."
-)
-
-type Config struct {
-	Logger     micrologger.Logger
-	FileSystem afero.Fs
-
-	Stderr io.Writer
-	Stdout io.Writer
-}
-
-func New(config Config) (*cobra.Command, error) {
-	if config.Logger == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
-	}
-	if config.FileSystem == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.FileSystem must not be empty", config)
-	}
-	if config.Stderr == nil {
-		config.Stderr = os.Stderr
-	}
-	if config.Stdout == nil {
-		config.Stdout = os.Stdout
-	}
-
-	f := &flag{}
-
-	r := &runner{
-		flag:   f,
-		logger: config.Logger,
-		stderr: config.Stderr,
-		stdout: config.Stdout,
-	}
-
-	c := &cobra.Command{
-		Use:   name,
-		Short: description,
-		Long:  description,
-		Args:  cobra.NoArgs,
-		RunE:  r.Run,
-	}
-
-	f.Init(c)
-
-	return c, nil
-}
