--- cmd/selfupdate/check/error.go
@@ -1,32 +1,0 @@
-package check
-
-import (
-	"github.com/giantswarm/microerror"
-)
-
-var invalidConfigError = &microerror.Error{
-	Kind: "invalidConfigError",
-}
-
-// IsInvalidConfig asserts invalidConfigError.
-func IsInvalidConfig(err error) bool {
-	return microerror.Cause(err) == invalidConfigError
-}
-
-var invalidFlagsError = &microerror.Error{
-	Kind: "invalidFlagsError",
-}
-
-// IsInvalidFlags asserts invalidFlagsError.
-func IsInvalidFlags(err error) bool {
-	return microerror.Cause(err) == invalidFlagsError
-}
-
-var updateCheckFailedError = &microerror.Error{
-	Kind: "updateCheckFailedError",
-}
-
-// IsUpdateCheckFailed asserts updateCheckFailedError.
-func IsUpdateCheckFailed(err error) bool {
-	return microerror.Cause(err) == updateCheckFailedError
-}
