--- cmd/selfupdate/check/flag.go
@@ -1,14 +1,0 @@
-package check
-
-import (
-	"github.com/spf13/cobra"
-)
-
-type flag struct {
-}
-
-func (f *flag) Init(cmd *cobra.Command) {}
-
-func (f *flag) Validate() error {
-	return nil
-}
